requirements.txt
matplotlib==3.7.1
numpy==1.23.5
pandas==2.0.2
Pillow==9.5.0
Pillow==10.0.0
streamlit==1.23.1
tensorflow
scikit-learn
